% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popReconstruct.R
\docType{package}
\name{popReconstruct}
\alias{popReconstruct}
\title{popReconstruct: A package for Bayesian population reconstruction.}
\description{
\pkg{popReconstruct} is a package for reconstructing populations of
the recent past.  It simultaneously estimates age-specific
population counts, fertility rates, mortality rates and net
international migration flows from fragmentary data, and
incorporates measurement error. Informative priors are required for
vital rates, migration rates, population counts at baseline, and
their respective measurement error variances.  Inference is based
on the joint posterior probability distribution which yields fully
probabilistic interval estimates.  A sample from this distribution
is drawn using a Markov chain Monte Carlo algorithm. The
popReconstruct package provides functions to implement Bayesian
population reconstruction for female-only populations \cite{Wheldon
et al., 2013, 2016} and two-sex populations \cite{Wheldon et al.,
2015}.
}
\section{Vignettes}{

See the vignette \pkg{burkina-faso-females} for a detailed illustration.
}

\references{
Wheldon, M. C., Raftery, A. E., Clark, S. J., & Gerland,
P. (2013). \dQuote{Reconstructing Past Populations With Uncertainty From
Fragmentary Data.} \emph{Journal of the American Statistical Association},
108(501), 96--110. \url{https://doi.org/10.1080/01621459.2012.737729}

Wheldon, M. C., Raftery, A. E., Clark, S. J., & Gerland,
P. (2015). \dQuote{Bayesian Reconstruction of Two-Sex Populations by Age:
Estimating Sex Ratios at Birth and Sex Ratios of Mortality.} \emph{Journal
of the Royal Statistical Society: Series A (Statistics in Society)},
178(4), 977--1007. \url{https://doi.org/10.1111/rssa.12104}

Wheldon, M. C., Raftery, A. E., Clark, S. J., & Gerland,
P. (2016). \dQuote{Bayesian population reconstruction of female
populations for less developed and more developed countries.}
\emph{Population Studies}, 70(1),
21--37. \url{https://doi.org/10.1080/00324728.2016.1139164}
}
